+++ Firestore/core/src/api/document_snapshot.cc
@@ -88,7 +88,10 @@ absl::optional<google_firestore_v1_Value> DocumentSnapshot::GetValue(
 bool operator==(const DocumentSnapshot& lhs, const DocumentSnapshot& rhs) {
   return lhs.firestore_ == rhs.firestore_ &&
          lhs.internal_key_ == rhs.internal_key_ &&
-         lhs.internal_document_ == rhs.internal_document_ &&
+         lhs.exists() == rhs.exists() &&
+         (lhs.exists() ? lhs.internal_document_->get().data() ==
+                             rhs.internal_document_->get().data()
+                       : true) &&
          lhs.metadata_ == rhs.metadata_;
 }